Moved the mouse events aggregation code to I_GetEvents
This seems more logical to me: motion events need to be aggregated
because of the way the game works not because of how mouses work or
something.
Also I needed the mouse input device class to correctly gather all the events to be able to get real-time information on mouse movements.